+++ CHANGELOG.md
@@ -15,6 +15,8 @@
 - add 'hive_configuration' member to IndeximaHook
 - change default time out
 - use a single cursor instance per hook process
+- set hive connection serialization encoding 'UTF-8'
+
 
 # 2.1.0 (2019-12-04)